+++ src/managers/UserLocationManager.m
@@ -22,7 +22,7 @@
 
 -(BOOL)addSubscriber:(NSString*)subscriberId;
 -(BOOL)removeSubscriber:(NSString*)subscriberId;
--(int)findSubscriber:(NSString*)subscriberId;
+-(NSUInteger)findSubscriber:(NSString*)subscriberId;
 -(void)removeAllSubscribers;
 
 
@@ -146,7 +146,7 @@ SYNTHESIZE_SINGLETON_FOR_CLASS(UserLocationManager);
 
 - (BOOL)hasSubscriber:(NSString*)subscriber{
 	
-	int index=[self findSubscriber:subscriber];
+	NSUInteger index=[self findSubscriber:subscriber];
 	
 	return index!=NSNotFound;
     
@@ -261,7 +261,7 @@ SYNTHESIZE_SINGLETON_FOR_CLASS(UserLocationManager);
 //
 -(BOOL)addSubscriber:(NSString*)subscriberId{
 	
-	int index=[self findSubscriber:subscriberId];
+	NSUInteger index=[self findSubscriber:subscriberId];
 	
 	if(index==NSNotFound){
 		
@@ -275,7 +275,7 @@ SYNTHESIZE_SINGLETON_FOR_CLASS(UserLocationManager);
 
 -(BOOL)removeSubscriber:(NSString*)subscriberId{
 	
-	int index=[self findSubscriber:subscriberId];
+	NSUInteger index=[self findSubscriber:subscriberId];
 	
 	if(index!=NSNotFound){
 		
@@ -293,9 +293,9 @@ SYNTHESIZE_SINGLETON_FOR_CLASS(UserLocationManager);
 }
 
 
--(int)findSubscriber:(NSString*)subscriberId{
+-(NSUInteger)findSubscriber:(NSString*)subscriberId{
 	
-	int index=[locationSubscribers indexOfObject:subscriberId];
+	NSUInteger index=[locationSubscribers indexOfObject:subscriberId];
 	
 	return index;
 }
